ingestion/src/requirements.txt
flask
psycopg2-binary
requests
faker
pandas
prometheus-client
watchdog